suggested changes implemented
